+++ darksite/cms/blog/test/factories.py
@@ -1,7 +1,9 @@
 import factory
 
+from test_utils import CleanAndSaveFactoryMixin
 
-class PostFactory(factory.DjangoModelFactory):
+
+class PostFactory(CleanAndSaveFactoryMixin, factory.DjangoModelFactory):
     """
     Factory for generating ``Post`` instances for testing.
     """
@@ -10,24 +12,3 @@ class PostFactory(factory.DjangoModelFactory):
 
     class Meta:
         model = 'blog.Post'
-
-    @classmethod
-    def _create(cls, model_class, *args, **kwargs):
-        """
-        Create a model instance, clean it, and save it.
-
-        Args:
-            model_class:
-                The model class to create an instance of.
-            *args:
-                The positional arguments to instantiate the model with.
-            **kwargs:
-                The keyword arguments to instantiate the model with.
-        Returns:
-            The created model instance.
-        """
-        instance = model_class(*args, **kwargs)
-        instance.clean()
-        instance.save()
-
-        return instance