+++ CHANGELOG.md
@@ -9,3 +9,4 @@
 * Added endpoint to fetch post list.
 * Added media resource model and capability to query for a media resource by ID.
 * Added model to track content of information panels shown on home screen and ability to query for information panels.
+* Added albums to list media resources.